ownkeysin refaktorointia
